re-re-factored the user class. Moved password hashing away from the data controller and into the CMS class. Added static parse fuinction to UserToken. Moved instance user addition function to CMS. Added isNoUsers function to determine if the user list is empty (may rename to not sound quite so stupid). Moved login and JWT creation away from data controller and into the CMS class. Implemented password change function.
This commit is realigning the purpose of the data controller. Initially, it was sharing some of the duties of the CMS, but it's becoming more clear that the data controller should solely be an intermediary between the CMS and the data location. It should not be doing any work involving token creation or password hashing or comparisons.